+++ buildSrc/src/main/kotlin/dependencies.kt
@@ -10,10 +10,10 @@ object rpgplugins {
 }
 
 object misc {
-    const val bstats = "org.bstats:bstats-bukkit:3.0.0"
-    const val annotations = "org.jetbrains:annotations:23.0.0"
-    const val serialization_hocon = "org.jetbrains.kotlinx:kotlinx-serialization-hocon:1.3.2"
-    const val mockk = "io.mockk:mockk:1.12.3"
+    const val bstats = "org.bstats:bstats-bukkit:3.0.2"
+    const val annotations = "org.jetbrains:annotations:24.1.0"
+    const val serialization_hocon = "org.jetbrains.kotlinx:kotlinx-serialization-hocon:1.6.1"
+    const val mockk = "io.mockk:mockk:1.13.8"
     const val adventure = "net.kyori:adventure-platform-bukkit:4.3.1"
 
     private const val commandapi_version = "9.2.0"
@@ -23,12 +23,12 @@ object misc {
 
 // Testing
 object junit {
-    private const val jupiter_version = "5.7.2"
-    const val jupiter = "org.junit.jupiter:junit-jupiter:$jupiter_version"
-    const val jupiter_params = "org.junit.jupiter:junit-jupiter-params:$jupiter_version"
+    const val bom = "org.junit:junit-bom:5.10.1"
+    const val jupiter = "org.junit.jupiter:junit-jupiter"
+    const val jupiter_params = "org.junit.jupiter:junit-jupiter-params"
 }
 
 object kotest {
-    private const val version = "5.2.3"
+    private const val version = "5.8.0"
     const val assertions = "io.kotest:kotest-assertions-core:$version"
 }