+++ mimic-bukkit/src/main/kotlin/MimicPlugin.kt
@@ -19,9 +19,6 @@
 
 package ru.endlesscode.mimic
 
-import dev.jorel.commandapi.CommandAPI
-import dev.jorel.commandapi.CommandAPIBukkitConfig
-import net.kyori.adventure.platform.bukkit.BukkitAudiences
 import org.bstats.bukkit.Metrics
 import org.bstats.charts.AdvancedPie
 import org.bstats.charts.SimplePie
@@ -32,7 +29,7 @@ import org.bukkit.plugin.java.JavaPlugin
 import ru.endlesscode.mimic.bukkit.loadAll
 import ru.endlesscode.mimic.bukkit.register
 import ru.endlesscode.mimic.classes.BukkitClassSystem
-import ru.endlesscode.mimic.command.registerCommand
+import ru.endlesscode.mimic.command.MimicCommands
 import ru.endlesscode.mimic.config.MimicConfig
 import ru.endlesscode.mimic.impl.battlelevels.BattleLevelsLevelSystem
 import ru.endlesscode.mimic.impl.customitems.CustomItemsRegistry
@@ -64,30 +61,21 @@ public class MimicPlugin : JavaPlugin() {
 
     private val config: MimicConfig by lazy { MimicConfig(this) }
     private val mimic: Mimic by lazy { MimicImpl(servicesManager, config) }
-    private var audiences: BukkitAudiences? = null
+    private val commands: MimicCommands by lazy { MimicCommands() }
 
     private inline val servicesManager get() = server.servicesManager
     private inline val pluginManager get() = server.pluginManager
 
     override fun onLoad() {
         Log.init(logger, debug = !isReleased)
-        CommandAPI.onLoad(CommandAPIBukkitConfig(this))
 
         servicesManager.register(mimic, this)
         hookDefaultServices()
     }
 
     override fun onEnable() {
-        CommandAPI.onEnable()
-        audiences = BukkitAudiences.create(this)
-
         if (isReleased) initMetrics()
-        registerCommand(
-            mimic = mimic,
-            config = config,
-            pluginFullName = description.fullName,
-            audiences = checkNotNull(audiences),
-        )
+        commands.register(this, mimic, config)
         pluginManager.registerEvents(ServicesRegistrationListener(servicesManager, mimic), this)
     }
 
@@ -193,9 +181,6 @@ public class MimicPlugin : JavaPlugin() {
     }
 
     override fun onDisable() {
-        CommandAPI.unregister("mimic")
-        CommandAPI.onDisable()
-        audiences?.close()
-        audiences = null
+        commands.unregister()
     }
 }